+++ algo/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## [0.1.1] - 2025-03-31
+## [0.1.1] - 2025-04-01
+
+### New
+
+* ExactSumSweep algorithm for eccentricities, radius, and diameter.
 
 ### Fixed